Fixed SYTEM_INFO structure for 64bit
This was only working in 32bit prior to this fix.

UInt64 happened to work in 32bit as two WORDs were missing making it look like everything was ok when reading dwNumberOfProcessors

DWORD_PTR should be a UInteger as it changes in size depending on architecture.

See https://blog.samphire.net/2017/01/22/windows-to-xojo-data-type-conversion/ for more information.